Removed legacy unused Cloud module. Renamed resource to infrastructure to match the frontend.
