--- src/app/business/cloud/cloud.component.html
@@ -1,11 +1,0 @@
-<p-tabView>
-    <p-tabPanel header="Service Registry">
-        <app-registry></app-registry>
-    </p-tabPanel>
-    <p-tabPanel header="Replication">
-        <app-replication></app-replication>
-    </p-tabPanel>
-    <p-tabPanel header="Migration">
-        <app-migration></app-migration>
-    </p-tabPanel>
-</p-tabView>
