--- src/app/business/cloud/cloud.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-cloud',
-  templateUrl: './cloud.component.html',
-  styleUrls: ['./cloud.component.scss']
-})
-export class CloudComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
