--- src/app/business/cloud/migration/migration.component.html
@@ -1,3 +1,0 @@
-<p>
-  migration works!
-</p>
