--- src/app/business/cloud/migration/migration.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-migration',
-  templateUrl: './migration.component.html',
-  styleUrls: ['./migration.component.scss']
-})
-export class MigrationComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
