--- src/app/business/cloud/replication/replication.component.html
@@ -1,29 +1,0 @@
-<!-- <div class="table-toolbar">
-    <div class="left">
-        <button pButton type="button" label="Create" (click)="showCreateUser()" class="ui-button-secondary"></button>
-        <button pButton type="button" label="Delete"></button>
-    </div>
-    <div class="right">
-        <div class="ui-inputgroup">
-            <input type="text" pInputText placeholder="search">
-            <button pButton type="button" icon="fa-check"></button>
-        </div>
-        <button class="" pButton type="button" icon="fa-check"></button>
-    </div>
-</div>
-
-<p-dataTable [value]="users" selectionMode="multiple" [showHeaderCheckbox]="true" [rows]="10" [paginator]="true" [pageLinks]="3" [rowsPerPageOptions]="[10,20,50,100]">
-    <p-column selectionMode="multiple" [style]="{'width': '60px'}"></p-column>
-    <p-column field="username" header="Username"></p-column>
-    <p-column field="status" header="Status"></p-column>
-    <p-column field="tenant" header="Tenant"></p-column>
-    <p-column field="role" header="Role"></p-column>
-    <p-column field="userid" header="User ID"></p-column>
-    <p-column header="Operation" [style]="{'width': '20%'}">
-        <ng-template pTemplate="body">
-            <a href="#">Modify</a>
-            <a href="#">Disable</a>
-            <a href="#">Delete</a>
-        </ng-template>
-    </p-column>
-</p-dataTable> -->
