--- src/app/business/cloud/replication/replication.component.ts
@@ -1,15 +1,0 @@
-import { Component, OnInit } from '@angular/core';
-
-@Component({
-  selector: 'app-replication',
-  templateUrl: './replication.component.html',
-  styleUrls: ['./replication.component.scss']
-})
-export class ReplicationComponent implements OnInit {
-
-  constructor() { }
-
-  ngOnInit() {
-  }
-
-}
